+++ example/lib/main.dart
@@ -30,24 +30,38 @@ class MyHomePage extends StatefulWidget {
 class MyHomePageState extends State<MyHomePage> {
   bool showEmojiKeyboard = false;
   final TextEditingController controller = TextEditingController();
+  int emojiReactionIndex = -1;
   bool showEmojiPopup = false;
   Offset popupPosition = Offset.zero;
+  List<String> emojiReactions = ['', ''];
 
   void backButtonFunctionality() {
+    if (showEmojiPopup) {
+      setState(() {
+        showEmojiPopup = false;
+      });
+      return;
+    }
     if (showEmojiKeyboard) {
       setState(() {
         showEmojiKeyboard = false;
       });
+      return;
+    }
+    if (Platform.isAndroid) {
+      SystemNavigator.pop();
     } else {
-      if (Platform.isAndroid) {
-        SystemNavigator.pop();
-      } else {
-        exit(0);
-      }
+      exit(0);
     }
   }
 
   void onTapEmojiField() {
+    emojiReactionIndex = -1;
+    if (showEmojiPopup) {
+      setState(() {
+        showEmojiPopup = false;
+      });
+    }
     if (!showEmojiKeyboard) {
       setState(() {
         showEmojiKeyboard = true;
@@ -55,18 +69,105 @@ class MyHomePageState extends State<MyHomePage> {
     }
   }
 
-  void onLongPressMessage(BuildContext context, LongPressStartDetails details) {
+  void onLongPressMessage(BuildContext context, LongPressStartDetails details, int messageIndex) {
+    emojiReactionIndex = messageIndex;
     setState(() {
       showEmojiPopup = true;
       popupPosition = details.globalPosition;
     });
   }
 
-  void closeEmojiPopup() {
+  void closeEmojiPopup(String? emoji) {
+    emojiReactionIndex = -1;
     setState(() {
       showEmojiPopup = false;
     });
   }
+
+  void handleEmojiPickerAction(EmojiPickerAction action) {
+    showEmojiPopup = false;
+    if (action is OutsideClicked) {
+      emojiReactionIndex = -1;
+    } else if (action is ButtonPressed) {
+      showEmojiKeyboard = true;
+    } else if (action is EmojiSelected) {
+      String newEmoji = action.emoji;
+      if (emojiReactions[emojiReactionIndex] == newEmoji) {
+        emojiReactions[emojiReactionIndex] = "";
+      } else {
+        emojiReactions[emojiReactionIndex] = action.emoji;
+      }
+      emojiReactionIndex = -1;
+    }
+    setState(() {});
+  }
+
+  void onActionEmojiChanged(String emoji) {
+    if (emojiReactionIndex != -1) {
+      emojiReactions[emojiReactionIndex] = emoji;
+      emojiReactionIndex = -1;
+      if (showEmojiKeyboard) {
+        showEmojiKeyboard = false;
+      }
+      setState(() {});
+    }
+  }
+
+  TextEditingController? getEmojiController() {
+    if (emojiReactionIndex != -1) {
+      return null;
+    } else {
+      return controller;
+    }
+  }
+
+  Widget messageWidget(int messageIndex, String messageText) {
+    return GestureDetector(
+      onLongPressStart: (details) => onLongPressMessage(context, details, messageIndex),
+      child: Stack(
+        children: [
+          Container(
+            color: emojiReactionIndex == messageIndex ? Colors.cyan[200] : Colors.transparent,
+            child: Container(
+              margin: const EdgeInsets.all(8.0),
+              padding: const EdgeInsets.all(12.0),
+              decoration: BoxDecoration(
+                color: messageIndex == 0 ? Colors.green[200] : Colors.blue[200],
+                borderRadius: BorderRadius.only(
+                  topLeft: messageIndex == 0 ? Radius.circular(30.0) : Radius.circular(0.0),
+                  topRight: messageIndex == 1 ? Radius.circular(30.0) : Radius.circular(0.0),
+                  bottomLeft: Radius.circular(30.0),
+                  bottomRight: Radius.circular(30.0),
+                ),
+              ),
+              child: Text(
+                messageText,
+                style: TextStyle(fontSize: 16.0),
+              ),
+            ),
+          ),
+          if (emojiReactions[messageIndex] != "")
+            Positioned(
+              bottom: 0,
+              right: 50,
+              child: Container(
+                padding: const EdgeInsets.all(4.0),
+                decoration: BoxDecoration(
+                  color: Colors.white,
+                  borderRadius: BorderRadius.circular(12.0),
+                ),
+                child: Text(
+                  emojiReactions[messageIndex],
+                  style: TextStyle(fontSize: 16.0),
+                ),
+              ),
+            ),
+        ],
+      ),
+    );
+  }
+
+  bool darkMode = false;
 
   @override
   Widget build(BuildContext context) {
@@ -88,68 +189,38 @@ class MyHomePageState extends State<MyHomePage> {
                 Expanded(
                   child: Container(),
                 ),
-                GestureDetector(
-                  onLongPressStart: (details) => onLongPressMessage(context, details),
-              child: Container(
-                margin: const EdgeInsets.all(8.0),
-                padding: const EdgeInsets.all(12.0),
-                decoration: BoxDecoration(
-                      color: Colors.green[200],
-                  borderRadius: BorderRadius.circular(8.0),
-                ),
-                child: const Text(
-                      'This is an example message. Long press to do an emoji reaction on this message!',
-                  style: TextStyle(fontSize: 16.0),
-                ),
-              ),
-            ),
-                GestureDetector(
-                  onLongPressStart: (details) => onLongPressMessage(context, details),
-              child: Container(
-                    margin: const EdgeInsets.all(8.0),
-                    padding: const EdgeInsets.all(12.0),
-                    decoration: BoxDecoration(
-                      color: Colors.blue[200],
-                      borderRadius: BorderRadius.circular(8.0),
-                    ),
-                    child: const Text(
-                      'Or press the Textfield below to start typing with the emoji keyboard!',
-                      style: TextStyle(fontSize: 16.0),
-                    ),
-                ),
-              ),
-                Padding(
-                  padding: const EdgeInsets.only(bottom: 76.0),
-                  child: Container(
-                    alignment: Alignment.topCenter,
-                    padding: const EdgeInsets.all(6),
-                    child: TextFormField(
-                      onTap: () {
-                        onTapEmojiField();
-                      },
-                      controller: controller,
-                      decoration: const InputDecoration(border: OutlineInputBorder()),
-                      readOnly: true,
-                      showCursor: true,
-              ),
-            ),
+                messageWidget(0, 'This is an example message. Long press to do an emoji reaction on this message!'),
+                messageWidget(1, 'Or press the Textfield below to start typing with the emoji keyboard!'),
+                Container(
+                  alignment: Alignment.topCenter,
+                  padding: const EdgeInsets.only(bottom: 76.0, left: 8, right: 8, top: 12),
+                  child: TextFormField(
+                    onTap: () {
+                      onTapEmojiField();
+                    },
+                    controller: controller,
+                    decoration: const InputDecoration(border: OutlineInputBorder()),
+                    readOnly: true,
+                    showCursor: true,
+                  ),
                 ),
                 Align(
                   alignment: Alignment.bottomCenter,
                   child: EmojiKeyboard(
-                    emojiController: controller,
+                    emojiController: getEmojiController(),
+                    onEmojiChanged: onActionEmojiChanged,
                     emojiKeyboardHeight: 440,
                     showEmojiKeyboard: showEmojiKeyboard,
-                    darkMode: true,
-        ),
-      ),
+                    darkMode: darkMode,
+                  ),
+                ),
               ],
             ),
             if (showEmojiPopup)
               EmojiKeyboardPopup(
                 position: popupPosition,
-                onClose: closeEmojiPopup,
-                darkMode: true,
+                onAction: handleEmojiPickerAction,
+                darkMode: darkMode,
               ),
           ],
         ),